requirements.txt
fastapi
uvicorn
python-decouple
openai
pinecone-client
sqlalchemy
psycopg2-binary
pydantic[email]
matplotlib
plotly
pandas
scipy
scikit-learn
passlib
bcrypt
python-jose
email-validator
python-multipart